Change return of div to double

+++ src/main/java/com/example/springdocker/model/MyMathCalc.java
@@ -10,7 +10,7 @@ public class MyMathCalc {
         return a * b;
     }
 
-    public int divide(int a, int b){
+    public double divide(int a, int b){
         return a / b;
     }
 }
